+++ src/components/audienceList/_AudienceList.scss
@@ -5,13 +5,15 @@
   width: 100%;
 
 
-  overflow-y: auto;
+
+  overflow-y: scroll;
 
   &::-webkit-scrollbar {
     width: 0 !important;
   }
   overflow: -moz-scrollbars-none;
   -ms-overflow-style: none;
+  overflow: scroll;
 }
 
 .audienceList__message {
