+++ src/components/plan/_Plan.scss
@@ -163,10 +163,14 @@
 }
 
 .showSearchMenu {
-  animation: showMenu 0.4s ease-out forwards;
+  visibility: visible;
+  opacity: 1;
+  transition: visibility 0s linear 0s, opacity 500ms;
 }
 
 .hideSearchMenu {
-  animation: hideMenu 0.4s ease-out forwards;
+  visibility: hidden;
+  opacity: 0;
+  transition: visibility 0s linear 300ms, opacity 500ms;
 }
