+++ src/components/popularLocations/_PopularLocations.scss
@@ -5,7 +5,6 @@
   padding-left: 10px;
 
 
-
   scrollbar-width: none;
 
   width: 100%;
