+++ src/components/searchBar/SearchBar.jsx
@@ -1,15 +1,15 @@
-import React, {useEffect} from "react";
+import React from "react";
 
 import searchBarDarkIcon from "../../images/searchBarDarkIcon.svg.svg";
 import searchBarLightIcon from "../../images/searchBarLightIcon.svg";
 
-const SearchBar = ({placeholder, props, searchQuery, isShowSearch}) => {
+const SearchBar = ({placeholder, data, searchQuery}) => {
     return (
         <div className="searchBar_packet">
             {/*<input className="openKeyboard" inputMode="text" disbabled/>*/}
-            <img src={isShowSearch ? searchBarLightIcon : searchBarDarkIcon} alt="search Icon"/>
+            <img src={searchBarLightIcon} alt="search Icon"/>
             {/*<span className="searchPsevdoInput_span">Поиск...</span>*/}
-            <input value={searchQuery} onChange={props} type="text" className="searchBar_input"
+            <input value={searchQuery} onChange={data} type="text" className="searchBar_input"
                    placeholder={placeholder} inputMode="text"/>
         </div>
     );
