+++ src/components/searchMenu/_SearchMenu.scss
@@ -8,8 +8,6 @@
 
   z-index: 20;
 
-  transform: translateX(-100%);
-
 }
 
 .SearchMenu__content {
@@ -18,11 +16,21 @@
   background-color: $c1-Background;
   height: 100%;
   padding-top: 15px;
+  padding-bottom: 10px;
   row-gap: 20px;
+  justify-content: space-between;
 
 }
 
+.SearchMenu__block {
+  display: flex;
+  flex-direction: column;
+  row-gap: 20px;
+  justify-content: space-between;
+}
+
 .SearchMenu__buttons {
+  justify-content: space-between;
   display: flex;
   align-items: center;
   column-gap: 20px;
@@ -31,6 +39,8 @@
     width: 20px;
     height: 20px;
   }
+
+  padding: 0 20px 0 10px;
 }
