+++ src/styles/main.scss
@@ -21,3 +21,4 @@
 @import "../components/popularLocations/PopularLocations";
 @import "../components/searchMenu/SearchMenu";
 @import "../components/audienceList/AudienceList";
+@import "../components/searchPsevdoInput/SearchPsevdoInput";
